+++ Server/src/main/java/model/context/spatial/Expanse.java
@@ -11,10 +11,10 @@ public class Expanse {
     private final Location bottomLeft;
 
     /** Breite des Kontextes. */
-    private final int width;
+    private final float width;
 
     /** Höhe des Kontextes. */
-    private final int height;
+    private final float height;
 
     /**
      * Erzeugt eine neue Instanz der räumlichen Ausdehnung eines Kontextes.
@@ -22,7 +22,7 @@ public class Expanse {
      * @param width Breite des Kontextes.
      * @param height Höhe des Kontextes.
      */
-    public Expanse(@NotNull final Location bottomLeft, final int width, final int height) {
+    public Expanse(@NotNull final Location bottomLeft, final float width, final float height) {
         this.bottomLeft = bottomLeft;
         this.width = width;
         this.height = height;
@@ -34,7 +34,7 @@ public class Expanse {
      * @param posY Zu überprüfende Y-Koordinate.
      * @return true, wenn sich die Koordinaten in der Ausdehnung befinden, sonst false.
      */
-    public boolean isIn(final int posX, final int posY) {
+    public boolean isIn(final float posX, final float posY) {
         return isAround(posX, posY, 0);
     }
 
@@ -45,7 +45,7 @@ public class Expanse {
      * @param distance Maximale Distanz.
      * @return true, wenn die Koordinaten die maximale Distanz zur Ausdehnung haben, sonst false.
      */
-    public boolean isAround(final int posX, final int posY, final int distance) {
+    public boolean isAround(final float posX, final float posY, final float distance) {
         return bottomLeft.getPosX() - distance <= posX && posX <= bottomLeft.getPosX() + distance + width
             && bottomLeft.getPosY() - distance <= posY && posY <= bottomLeft.getPosY() + distance + height;
     }
