+++ Server/src/main/java/model/context/spatial/MapUtils.java
@@ -55,12 +55,12 @@ public class MapUtils {
         return communicationMedia;
     }
 
-    public static int getWidth(@NotNull final TiledMap tiledMap) {
+    public static float getWidth(@NotNull final TiledMap tiledMap) {
         TiledMapTileLayer layer = (TiledMapTileLayer) tiledMap.getLayers().get(0);
         return layer.getWidth() * layer.getTileWidth();
     }
 
-    public static int getHeight(@NotNull final TiledMap tiledMap) {
+    public static float getHeight(@NotNull final TiledMap tiledMap) {
         TiledMapTileLayer layer = (TiledMapTileLayer) tiledMap.getLayers().get(0);
         return layer.getHeight() * layer.getTileHeight();
     }
@@ -87,12 +87,12 @@ public class MapUtils {
     }
      */
 
-    public static int getSpawnPosX(@NotNull final TiledMap tiledMap) {
-        return tiledMap.getProperties().get("spawnPosX", Integer.class);
+    public static float getSpawnPosX(@NotNull final TiledMap tiledMap) {
+        return tiledMap.getProperties().get("spawnPosX", Float.class);
     }
 
-    public static int getSpawnPosY(@NotNull final TiledMap tiledMap) {
-        return tiledMap.getProperties().get("spawnPosY", Integer.class);
+    public static float getSpawnPosY(@NotNull final TiledMap tiledMap) {
+        return tiledMap.getProperties().get("spawnPosY", Float.class);
     }
 
     /**
@@ -126,10 +126,10 @@ public class MapUtils {
         String areaName = contextRectangle.getName();
         CommunicationRegion communicationRegion = getCommunicationRegion(contextRectangle.getProperties());
         Set<CommunicationMedium> communicationMedia = getCommunicationMedia(contextRectangle.getProperties());
-        int posX = (int) contextRectangle.getRectangle().getX();
-        int posY = (int) contextRectangle.getRectangle().getY();
-        int width = (int) contextRectangle.getRectangle().getWidth();
-        int height = (int) contextRectangle.getRectangle().getHeight();
+        float posX = contextRectangle.getRectangle().getX();
+        float posY = contextRectangle.getRectangle().getY();
+        float width = contextRectangle.getRectangle().getWidth();
+        float height = contextRectangle.getRectangle().getHeight();
 
         Expanse expanse = new Expanse(new Location(room, posX, posY), width, height);
         // Ermittle den übergeordneten Kontext. Da die Quadrate absteigend ihrer Grö0e sortiert eingefügt werden,
